Fix issue #10: Top-level functions and variables not visible in local scopes.
Problem

In both Pyonic 2 & 3, functions that are defined and variables that are assigned at the top-level of the interpreter are not visible as globals within other functions (or even themselves -- no recursion). A simple example:

def foo():
  print("foo")

def bar():
  foo()
  print("bar")

bar()

The call to foo() from bar() raises NameError with the message "global name 'foo' is not defined" in Python 2.7 and "name 'foo' is not defined" in Python 3.6.

Solution

The Python 3.11 documentation says this about the built-in exec() function: "If only globals is provided, it must be a dictionary (and not a subclass of dictionary), which will be used for both the global and the local variables ... Remember that at the module level, globals and locals are the same dictionary. If exec gets two separate objects as globals and locals, the code will be executed as if it were embedded in a class definition." <https://docs.python.org/3/library/functions.html#exec>

So by passing two different objects as globals and locals to exec(), the Pyonic code was incorrectly asking Python to run it as part of a class definition.  globals() and locals() return the same object in the top-level module.  Therefore, only one dictionary should be passed to exec().